Reduces size on move index to 8bit.

+++ Controller/Searcher/BreadthFirstCubeSearcher.h
@@ -25,9 +25,9 @@ namespace busybin
   {
     struct node
     {
-      uint16_t moveInd;
-      node*    parent;
-      uint8_t  depth;
+      uint8_t moveInd;
+      node*   parent;
+      uint8_t depth;
     };
 
   public:
